Add specialized methods for reading and writing kernel version to log file
We write the kernel version several times per transaction, once per command to be precise, which is unnecessary since mixing them in a single transaction makes no sense.

Here we introduce specific methods to read and write the version, so that channel implementations can choose to ignore them. We have to do this to support rolling upgrades, without duplicating all the writing and reading code.