Extract constants when possible
A common pattern is doing things like `WHERE datetime(#p) < n.p` which will
evaluate the datetime function for reach row even though it is really constant
for the query. For temporal functions this is quite costly. This PR introduces
a rewriter that finds values we know are constant for the queries and make sure
we only evaluate them once.