Less blocking of countstore writers on full cache
When the cache got full several writers could get in and try to take
the exclusive lock. This meant that no writes could continue even
after the critical section (when having downgraded to shared lock)
because the next writer was waiting for the exclusive lock and fair
locking is used.
Now only one thread wais for the exclusive locks and the others will try
to take the shared lock instead meaning that they will be able to
coninue when the critical section has passed.

Co-authored-by: Valdemar Roxling <valdemar.roxling@neotechnology.com>