Split leaf and internal node logic in TreeNode ()
This is a refactoring of TreeNode and related classes with the goal to separate code responsible for leaf and internal nodes encoding. With multiversion tree work leaf nodes behave quite different to regular tree and there will be more diversion, i.e. I need extra data in header only in leaves. With such separation I can easily change leaf node behaviour without affecting internal nodes.

Interface of the TreeNode class is left intact and it works as a boundary of this refactoring. Its methods now delegate either to LeafNodeBehaviour or to InternalNodeBehaviour; childs like TreeNodeFixedSize and TreeNodeFixedSizeVersioned responsible only for instantiation of implementations of those behaviours.

Code that is common to leaf and internal nodes converted to static functions and moved to DynamicSizeUtil or TreeNodeUtil. Same for operations with the header. 

At TreeNode level I added keyAtLeaf and keyAtInternal methods, and use them wherever type parameter was constant (basically everywhere except SeekCursor). In addition I added specialised versions of keysearch.