Fixes an inconsistency issue recovering created indexes
Basically a race in recovery where an index, created and
fully populated after the last checkpoint, could not be repopulated
and not be updated during recovery. This would result in an ONLINE
index which was out of date with the graph, resulting in an
inconsistent database.

See more details in IndexCreationRecoveryIT