Fix bug with missing Eager in before Subqueries
For DELETE conflitcs, we tracked the last plans to reference a variable.
That, in combination with `foldPlan` and snapshots did not work well
with Apply variants (e.g. SubqueryForeach), when we have "last" references
both on the LHS and the RHS of the Apply.

This commit changes it such that we tack _all_ references to a variable instead.
This fixes the bug, but it also introduces more EagernessReasons which can be quite verbose.