Sub-query error in pipelined runtime
In pipelined runtime when having an aggregation on the lhs of an Apply under a CartesianProduct we used the wrong slot configuration ending up an error when copying data.

Fixes: #13148
changelog: Fix `A bug has occurred in the morsel runtime: The target morsel cannot hold the data to copy` error. See Github issue: neo4j/neo4j.