test: fix FrekiDatabaseUpgradeTransactionIT to requie a valid version to upgrade to
The point of these tests are to test rolling upgrade between two actual
kernel versions. For freki, this would have to be the first kernel
version supporting the finalised V1, and the kernel version after.

This test won't be semantically correct until that point. When written,
it was a quick hack, generating a near empty database, and coercing the
versions. This was to ensure the expected code paths were correctly
utilised. The aim was always to move to a zipped store, like its record
format parent; however, that won't be possible until we finalise V1 of
the format.

This was before the work surrounding rolling upgrade that enabled
wider use of GLORIOUS_FUTURE. UpgradeToFutureVersionIT now covers the
testing for rolling from whatever LATEST is, to GLORIOUS_FUTURE; testing
all these code paths. Keeping {,Freki}DatabaseUpgradeTransactionIT
to semantically checking expected rolling between real kernel
versions. Thus FrekiDatabaseUpgradeTransactionIT is not needed
in the same capacity as it once was, as it is being tested by
UpgradeToFutureVersionIT.

This is why we can temporarily disable this test; which was need due to
the extra checks around freki and kernel versions. The test will now
fail when selecting older kernel versions that cannot actually support
freki. Thus we should keep this disabled until the point we have two
kernel versions supporting a stable freki V1; which we will also move to
a zipped store.