Define which system graph components have versions, and tombstone Fabric ()
Some components use actual version numbers, others can just detect if they can perform an upgrade. It's useful to distinguish one from the other, so we add the SystemGraphComponentWithVersion interface.

Additionally, the FabricSystemGraphComponent claimed to provide a version number, but didn't really because it writes nothing to the system graph. It has been updated to write its version number to the Version node in order to tombstone it - mark it as unambiguously upgraded to a final version, never to change.