Improving Fabric error message after leader switch
This change provides a better error message for the situation when there is a writing fabric/query-router transaction in progress and a leader switch occurs. An error saying "Writing to more than one database per transaction is not allowed" would have been returned if the user continued to write to the same database. The reason was that even thought it was the same database, the leader location changed, so it was seen as a different target location for the write. This was quite confusing for the users.

I tried to create a stable cluster test for this situation and failed misserably.