Freed IDs enter the cache directly w/o marking or scanning
This heavily reduces I/O, both from avoiding marking and from
avoiding scanning to let those IDs enter the cache.
So now the states for an ID becomes:

- used
- deleted
- either of:
  - *cached* (w/o getting marked as either free or reserved)
  - free (and later reserved as it's found in a scan)

So an ID in the cache is either marked as deleted+free+reserved,
or more typically _only_ deleted. IDs that gets e.g. rolled back
or on cleared cache gets the free mark added and its reserved mark
removed to restore the correct state for a scan later.