Use List for `openSequences`
This changes such that we use a `List` instead of a `Seq`
and prepend instead of append. This improves performance.

Also pass around more `Refs`, which means we need to
create less often new `Ref` objects.

Some tests had to be adapted because the
list of candidates is now built in reverse and
if there were candidates with the same cardinality,
a different one would get picked now.