Write instance versions to system graph ()
Adding a maintenance job that will periodically see if it needs to write the instance component versions found in discovery down to the instance in the system graph. It also checks against the current general Version node to see if the new versions are compatible, and logs problems if not.