Defaults import --strict to false
There have been many reports of importer being a lot slower
since strict option was introduced (and it's true by default).
Not a little, but a couple of times slower in total.
The main reason is that the page cache used to do this lookup
is the same one that the importer uses, and it's typically 1GiB in size.