Changing the strategy how multiple errors are handled in query router transactions
Before, if there were multiple errors during commit, rollback or a similar operation, the errors were added as supressed exception to a generic exception. Now, one of the errors is chosen as the main one and the remaining ones are added as supressed ones.
This change allows us to better preserve the original exceptions and avoid less informative generic ones.

One illustrative example where the new behaviour is useful is when databses are shut down. We used to produce a generic exception with multiple supressed "database unavailable" exceptions.