Make DBMS upgrade procedures check component versions per instance ()
Using the newly available component versions for each instance, which are written into the system database, the procedures can verify that they all have the same versions before allowing an upgrade to proceed. They can also ensure that the converged versions are an apparently compatible upgrade from the current agreed versions. The status procedures will report a new status of CANNOT_UPGRADE if there is a conflict / lack of convergence.

In addition, we increase the frequency of running TopologyMaintenance, so that instance versions get written down more promptly, avoiding upgrades having to wait five minutes for the component versions to get written before an upgrade is permitted.