fix: consistency checker report file
According to the documentation of the consistency check options:

      --report-path=<path>   Path to where a consistency report will be written. Interpreted as a directory, unless it
                               has an extension of '.report'
                               Default: .

Unfortunately when provided a specific path to a file, with the
extension, it was still interpreted as a directory.

For example:
        --report-path="consistency-check.report"

Would generate:
        ./consistency-check.report/inconsistencies-2023-05-05.16.27.49.report

The culprit was Path::endsWith, according to the documentation:
> On UNIX for example, the path ends with "foo/bar" and "bar".
> It does not end with "r" or "/bar".

Thus switched to org.apache.commons.ioFilenameUtils::isExtension to
ensure extensions are properly checked.

Tests were added to explicitly check the ConsistencyCheckService
generated the expected report file path.