Support transaction memory tracking in parallel runtime
Transaction memory tracking for parallel runtime is supported by injecting
memory trackers that delegates to the execution context memory tracker of
the worker thread.

Since we do not controll which individual worker is doing corresponding
allocate and release calls on shared data structure, a special version
of memory tracker is needed for execution contexts that is also allowed
to go negative on allocated heap bytes, with the only requirement that the
combination of all execution context memory tracker allocations is going
to sum up correctly in the transaction memory pool.
This ExecutionContextMemoryTracker is made adaptive to dynamically adjust
the reserve/release size toward the underlying transaction memory pool
in order to reduce contention while also trying to limit the upfront memory
grab size.

For shared data structures we use a ConcurrentScopedMemoryTracker that
limits contention using LongAdder.

Query memory tracking per operator that is used with PROFILE mode
is not yet supported.