refactor: add fallback to latest runtime version internal setting for tests
Introduce internal setting to switch to fallback to the latest runtime
version, rather than use the otherwise semantically chosen value for 5.0

This is needed when specifically disabling system graph initialisation,
but still needing to test features newer than 5.0.

It is not ideal to modify production code for test purposes, but adds an
automatic enforcement where appropriate, protecting current and future
tests. This minimised changes to production code, adding a single config
lookup on DBMS startup.