test: fallback to latest runtime when disabling system graph initialisation
There are some tests/benchmarks that require disabling the system graph
initialisation; in doing so there is no system graph in which to query.
When the runtime is queried from a non-existent graph, it would fallback
to 5.0 (which is the semantic choice).

Of those tests/benchmarks, they may require features that are blocked
behind kernel versions; this could be any feature, the case that
triggered this, was freki.

Using the previously defined internal setting whenever this
initialisation is explicitly disabled, prevents these tests from hitting
version based assertions and exceptions, that otherwise wouldn't have
occured in the presence of a system graph.

As this modifies the provided config, the non-config default was
removed. Considering the couple of places it was used had access to the
config, it was no needed.