Fix flaky mismatch cursor count failure in parallel TCK-tests
Fixes a race condition causing incorrect reporting of counts to the
page cache tracer when closing a transaction, after running an embedded
query with the parallel runtime and using the Core API on returned entities
within a result transformer.

The problem was that the ValueMapper used in ResultSubscriber was exposing
the internal transaction directly to returned entities, which could lead
to uses of ambient cursors updating non-threadsafe counts while execution
contexts were merging their reports back to the transaction on query close.

This is solved by providing the subscriber with an ExecutionContextValueMapper
based on the execution context of the resources used for query intialization.
These resources now have to be closed from the query calling thread after
merging of worker execution contexts is completed.