Fixes an issue where GBPTree would not decrease its depth
If all these conditions were met:

- dynamic layout
- a root with one key, i.e. with two children
- replacing a value of one of the entries in the left child
- the replaced value is smaller than its existing value
- the reduction in size would cause underflow
- the combined contents of left child (w/ the smaller value) plus the
  right child would fit into one page

Then the two children would be merged together, but the tree depth
wouldn't shrink, leaving the tree looking like an internal root
with no keys and its only child having some entries in it.
Even if all those entries would be removed, the tree wouldn't shrink.

Tho this was not a problem for single-root trees, more of an
inefficient structure - for multi-root it would cause the problem
where if such a data tree would be deleted, the child would not be
placed on the freelist and therefore its page lost/unused forever...
something that the gbptree structure checker picked up on.