Locks token index earlier for scanning
When doing a store scan for index population the token index,
if present, was locked during the time the store scan ran.
This prevented the token index to be deleted while running
the scan. However, this was too late because if that index
was deleted after the decision to use it was made and before
the run started (though a small window of time), the population
would fail and unnecessarily put the index(es) participating in
that population in FAILED state.

Now the token index is locked as part of making the decision whether
to use it for the scan or not, avoiding the inconvenience of
failing populations using it.