Adding licenses
Adding licenses to some of the files moved in the previous commit.
Changes to a file must not be too large in order for Git to consider
them a move and not loose the file's history. Linces were ommited
from some of the small files in the previous commit for this reason.