Remove QueryGraph.writeOnly, simplify code, accept regressions.
`QueryGraph.writeOnly` was quite certainly incomplete, missing reads in mutating patterns like
* HasLabel
* Labels(n)

Instead of trying to find the missing cases, this removes this optimization altogether.
This enables us to clean up some now identical methods in IR eagerness.

It leads to a few harmless regressions, where we now start a plan with

```
.eager
.create
.argument
```

This is harmless since we are only eagerizing a single row.