Plan selective path patterns
expandSolverStep plans selective path patterns using statefulShortestPath. For that it constructs an NFA. To do so, ConvertToNFA takes the pattern and a direction. It inlines all predicates possible (those that depend on available symbols and the variable at hand) and returns the other predicates as non inlined predicates. Predicates depending on no path variables are planned before statefulShortestPath.

For now, we can only start from one side of a pattern and work our way to the other side. That is, if the end node is already bound, a hidden selection is introduced for SPPs

Currently not supported:
- Interior variables may not have an equijoin, that is, they must not repeat the name of other variables in the pattern.
- Predicates inside QPPs referring to more than one variable inside the QPP.
- Var-length relationships

Co-authored-by: Satia Herfert <satiaherfert@gmx.de>