Move predicates on boundary nodes to post-filter position
A predicate on a boundary node in a path pattern with a selector
can be moved from pre-filter position (inside ParenthesizedPath)
to post-filter position (in top-level MATCH).

This is semantically equivalent, since all selectors apply per
partition of (left boundary node, right boundary node) tuple,
so any predicates that only depend on those 2 variables can be applied
either as a pre- or as post-filter.

Having them in post-filter position will allow the planner to plan these
before the invocation of the shortest algorithm, and choose better leaf plans.