Fixing an issue when 'database not found' error is incorrectly thrown
There is a small time window between the routing decision being taken and a local database
transaction starting when the local database can be migrated from the server instance.
This situation used to lead to 'database not found' error which is a bit confusing, because
the database exists in the cluster. Moreover, this error is classified as a client one. 'Database unavailable'
transient error is thrown now in this situation which apart from being less confusing has the desired properties
associated with transient errors like driver retrying.