Refactor procedure/kernel interaction
We were inconsistently handling injections when using parallel runtime. In some
cases, e.g., `@Context Transaction tx`, we were providing a dumbed down version
whereas in the case of `InternalTransaction` and `KernelTransaction` we were
failing hard when trying to inject the instance. This is not only inconsistent
but also makes it harder for users since not all procedures in the procedure
file actually use the injected instance.