Remove NFA.State.VarName, but add singletonVariables to StatefulShortestPath ()
Before, each state would keep track whether its variable is a group or a singleton variable. This would be used by runtime to determine how to project the node/rel ids into the outgoing row. This information was partially overlapping with

```
StatefulShortestPath.nodeVariableGroupings
StatefulShortestPath.relationshipVariableGroupings
```

Now, we add `StatefulShortestPath.singletonVariables`, consolidating all necessary information into StatefulShortestPath. The NFA thus does not need to track this any longer.

It is still encouraged to generate row-setter-lambda functions and put those into states of the compiled version of the NFA, but the logical NFA does not need to have this information.

Co-authored-by: Satia Herfert <satiaherfert@gmx.de>