Add procedure to set vector property
Cypher: LIST<FLOAT> -->   Java: List<Double>
                    --> Values: DoubleArray
                    --> stored: double[]

However, float[] is usually what embeddings are generated and
manipulated as. The Lucene requires float[], thus we convert when
indexing.

So generated embeddings would often be stored in twise the space they
need in the datebase. Until we have an explicit type, this procedure
will convert and set the input as a float[] internally.