Fix flaky test in ReadAndDeleteTransactionConflictIT ()
When a reader of nodes is competing with a deleter of nodes, there is a
possibility for a race relating to dynamic labels.

In the test `ReadAndDeleteTransactionConflictIT.nodesThatAreConcurrentlyDeletedWhileStreamingResultThroughBoltMustBeIgnored`
this race sometimes shows itself as an exception:

    org.neo4j.kernel.impl.store.InvalidRecordException: DynamicRecord[309,used=false,(23),type=-1,data=byte[size=23],start=true,next=-1,created=false] not in use

During a normal execution, nodes are read as below, where certain nodes
are "heavy" while others remain "light" as seen from the print below.

    node<730> (4:ef59c6ac-aa4b-4e14-89b8-b571bf23f4cb:730): [A, L, H, C, D, P, K, G, E, B, I, L6, F, Q, J, O] [[]]
    node<731> (4:ef59c6ac-aa4b-4e14-89b8-b571bf23f4cb:731): [] [[]]

One issue that can occur, is the following:

1. The reader creates a "light" node.
2. The deleter removes nodes, and the associated labels.
3. The reader attempts to make the "light" node "heavy", to populate the
   labels on the node.

Currently, this results in a crash when the reader attempts to read a
`DynamicRecord`s with `inUse=false`. This is due to the node reading the
label with `ReadMode.NORMAL`, which enforces the `inUse` flag to be
true.

This patch attempts to address the issue, by being more lenient when
loading the DynamicLabel by using `ReadMode.ALWAYS` instead. The
motivation is that since labels can be set to `inUse=false` at any time,
any place where it actually matters should already use
`node.getUsedDynamicLabels()` over `node.getDynamicLabels()`. The same
choice can also be found in `RecordPropertyCursor` where the same
lenient `ReadMode.ALWAYS` is used.

Furthermore, the patch removes an excessive `isLight` check that is done
before `node.ensureHeavy()`, since said function itself makes the check.