Respect skip/limit from index constraints
Though SKIP and LIMIT (with YIELD) can be used to ignore or reduce the
output; fulltext allows for skip/limit to be provided via options to
minimise the amount of work that actually needs to be done.

As vectors work in a similar way, this lays the foundation for them to
be used in the future.

Co-authored-by: Matthew Parnell <matthew.parnell@neo4j.com>