5.x Use new kernel checkpoint API ()
The main purpose of this PR was to make sure that the checkpoint entry after a full store copy is written to the last transaction in the store. This was not the case before where we would just write the entry on the first pulled transaction. With CDC enabled, all transactions will be pulled in a full store copy. So we must write the checkpoint tx to the last committed transaction in the store, otherwise we’d reapply all those transactions.

While doing this we realized that after a full store copy, we’d always pull the last committed transaction again, applying it twice and potentially corrupting the store.

So the biggest change in this PR is that we don’t pull the last committed transaction anymore. However, this transaction had a purpose. It was used to get the raft index of the last committed transaction in the store. Without this extra transaction we had to come up with another way to link the last committed transaction to its corresponding raft index. Kernel helped us store this information in each checkpoint entry instead, so it can now be accessed from the metadata store.

Changes to RequiredTransactions and TxPullRequestContext:
As we’ll never pull the last committed transaction inclusively anymore, much of the logic in here could be removed. We also squashed RequiredTransactions and TxPullRequestContext together, as they were doing essentially the same things.

Previously we had to first check if there were any transaction logs. In that case, the tx pulling started from the lastTxInLogs + 1. If there weren’t any transaction logs we’d get the lastCommittedTxid from the MetaDataStore and pull from that transaction inclusively. If no other remote had that transaction in their logs (because all members were restored with a backup without logs), it would “fallback” to pull from lastCommittedTxId + 1.

Note that we, since the kernel version work, don’t allow pulling transactions without transaction logs anymore.

Changes to TxCatchupSession:
The last checkpoint transaction is now sent over with the PrepareStoreCopyResponse, so that the checkpoint entry can be written to it. So a new catchup protocol version was introduced. The transaction will also include the kernel version which gets written down in the checkpoint entry.

We also did some preparatory work for CDC in this area. After a full store copy, the transaction log will be created only once the first transaction is received. If no transaction is pulled, it’s created right before writing the checkpoint entry and closing the session. This is important because, when creating the log files, we need to set a lastCommittedTxId in the header. This represents the first transaction in the log - 1. Previously we’d create the log file with the last committed transaction in the store but with CDC this is not true anymore.

Getting checkpoint tx from earlier version remotes
It’s only possible to get the checkpoint tx in the new catchup version. So when getting the store listing from a legacy node, the checkpoint tx won’t be included in the PrepareStoreCopyResponse. In this case, the checkpoint tx will be pulled with the heap tx puller instead.

Co-authored-by: tselmegbaasan <tselmeg.baasan@neo4j.com>