Auto enable free servers ()
This PR adds functionality to have servers automatically be enabled when they are discovered.

A new initial config setting initial.dbms.cluster.automatically_enable_free_servers is introduced; defaulting to false. We then introduce a auto_enable_free_servers property on the topology graph config node, which can be modified at runtime using the dbms.cluster.setAutomaticallyEnableFreeServers procedure. The current value can be viewed using the dbms.showTopologyGraphConfig() procedure.

The AutoEnableServerChecker is implemented as a TopologyMaintenanceCommand that runs every 5s. It reads the auto_enable_free_servers property from the model and if set true, compares the servers it can see in discovery and in the model, for any in discovery that are not in the model it writes an instance node to the topology graph. It uses the serverConfig information found in discovery.