5.x Pull historical transactions (CDC) ()
With CDC we need to make sure that each cluster member always has all transactions in its log. This is not the case today because when a new member joins for example, it performs a full store copy. Also, when a member falls behind too much in raft, it will perform a full store copy. A full store copy starts by downloading the store files. Often there are more transactions applied while the store files are downloaded. In these cases, those recent transactions will be pulled as a finishing action in the full store copy. But the historical transactions that are already covered by the store files will never be pulled. We need to make it possible to pull these historical transactions, so that new members and catching up members get all transactions.

We've added a new filed field `pullHistoricalTxs` in `RawTxPullRequest`, which, if true, tells the server to send all available transactions. Note that this is only needed when the tx pulling is preceded by a full store copy.
For all other cases, there is already a transaction log with the historical transactions, and the tx pulling is only used for fetching the missing transactions.

Apart from the cluster-internal catchups, the tx puller is used by the Backup admin command. For backups, the historical transactions should never be pulled, regardless of if CDC is enabled or not.

### Backwards compatibility
When talking to a legacy member which doesn't have a notion of the new `pullHistoricalTxs` field, the client will do a normal full store copy, namely pull all transactions from the last transaction committed in its store files.

Co-authored-by: fabiim <dev.fbotelho@gmail.com>