Fix enrichment object to allow for concurrent serialization ()
Problem was found in cluster where the tx log was corrupted on the leader (enrichment command was truncated or read was aborted early due to a corrupted buffer sizing). It was observed that the leader could write the enrichment command as follows:

write the command to the raft log
start writing the same command to follower 1 across the network
start writing the same command to follower 2 across the network
follower 1's write completed quickly whilst follower 2's write proceeded slowly
follower 1 ack'd the write BEFORE follower 2 had start/completed it's write across the network
leader would write the same command to the tx log.
In the very rare cases of this occurring, it could be possible for the Enrichment objects serialize method to be called concurrently. The WriteEnrichmentChannel serialize method however was NOT thread-safe and could be the source of the corruption.

Fix is to use a slice of the WriteEnrichmentChannel's internal ByteBuffers so that any position/limit markers would be unique for each serialize call and hence avoid any concurrent modifications.