ReadFinder: Use variable name for type lookup
The LogicalPlans can currently contain variables with
incorrect input positions. This is an interim state
until https://trello.com/c/c3fTtlQD is fixed.

When using expressions, we can therefore be over-eager
on such variables. When using the name, we find the correct
variables. This is safe because LP Eagerness runs after
the Namespacer.

Test that would fail otherwise:
"should not be eager when deleting nodes in subqueries that cannot overlap with unstable node from outside subquery, when solved with LabelScan that excludes deleted node"