Reduce noise level from ArchiveProgressPrinter ()
.. by allowing it to print **either** on a new percentage point, or when
60 seconds has elapsed without a print.

Since the scheduling check is only performed in `addBytes`, we can
without any change of functionality move the scheduling from the
background thread into the `addBytes`. This allows us to fully remove
the, now unused, background thread.

In addition, the commit allows the ArchiveProgressPrinter to be tested
with a controllable time source, which allows testing how noisy the
printer actually is.