Add a loop aware logger for Raft binding ()
Raft binding runs in a loop, and sometimes retries the same loop repeatedly on a short interval, which can spam the logs.

The loop aware publisher will ignore loops where all the messages are the same as the last loop, but if any of the messages in a loop are different, it will log all the messages. This will affect their timestamp, and the order relative to log messages from other loggers in that case, but will preserve the order of messages within the logger.