GlobalProcedures produces immutable snapshots ()
.. of the available procedures. These are updated on every modification to the ProcedureRegistry (~56 of those during a normal startup), and are thus fairly costly. To remedy this, we introduce a new bulk registration mode which reduces the number of snapshots produced to only 2 (one for registrations, and one for the `start`-method.

This commit will introduce a "breaking" change for apoc-extended, in the case that the user registers a procedure with apoc.custom.declareProcedure and attempts to use it in the same transaction. Where it would previously work(-ish) it will now fail due to the fact that newly registered procedures now require a new transaction to become visible.