Fix FulltextIndexCapability
The value and category checks were not use (to my knowledge); thus, this
wasn't caught before.

The old behaviour effectively required that every value had to be text,
text-array, or no-value; however, they couldn't all be no-value.

The actual behaviour should be simpler; and simply only one value that
is text or text-value needs to exist.

This was likely due to a miscommunication when implementing this for
fulltext when introduced.

Whilst here I noted that an all entries scan is a valid query for
fulltext. (also likely not checked)