Implement count indexed entities
Currently only checks entity is in the index; it does not check the
value is the same.

Couldn't find a quick solution in current time contraints to extract the
needed information from the index.

Next steps: investigate using LeafReader::getFloatVectorValues with
something like DocValuesCollector.

An alternative if that doesn't pan out might be a slightly more
expensive check of k-ANN with k=1, filter=getById, (score-1) < epsilon