Don't rewrite grouping columns with variables in leverageOrder
Doing so might introduce a cycle in names, which would fail in
InsertCachedProperties.

When we have a clause like:

```
RETURN DISTINCT a AS a, a AS b ORDER BY b`
```

We might push sorting below `Distinct`, which would also
introduce `a AS b` projection, e.g:

```
...
.sort("b ASC")
.projection("a AS b")
.allNodeScan("a")
```

After that we would previously rewrite `a AS a` column in DISTINCT into
`b AS a`, thus creating a cycle.

Fixes #13232