Move ReconciledTransactionTracker to enterprise
This interface had two implementations, from which one was not used at all. That one is removed.
The rest is moved from the module neo4j-bolt (in community) to neo4j-dbms-enterprise. The interface and the implementation is used nowhere in community or in bolt.
Once it might have to do with bolt, but at this point this is mainly important for enterprise/reconciler as well as the wait procedure only.