Mvcc countstore recovery after compaction ()
During checkpoint It can happen that visibility horizon went past the transaction id that marks checkpoint, by the time it come to countstore.
This means that reverse and forward recovery will go through compacted counter which should be carefully preserved.

For example, we can end in the state like this:

 TxId:       9 | 10 | 11 | 12 | 13 | 14    | 15 | 16 | 17 | 18 | 19
 Store:                            | +1000 | +1 |
 Commands:     | +2 | -3 | +1 | +6 | -22   | +1 | -1 | +1 | -1 | ...
Checkpoint is at 9, reverse recovery reverses up to 10, compacted counter is at tx 14.

To preserve compacted counter and properly perform reverse-forward recovery cycle we need:

remove or set to zero any delta that is after tx 14.
for any transaction before tx 14, its negated delta should be applied to the delta at tx 14.
To implement this we introduce new operation on the tree updateCeilingValue. It looks for the key that is equal or greater then specified key, and if it is found, applies desired change to the value of that key. It also have upper boundary for search to distinguish sequences of deltas related to different counters.