Do not merge CREATE clauses if they have dependencies
We merge two consecutive `CREATE` statement to be able to create them together. However, we may not do that if the second statement depends on the first one to have taken place.
Currently, this can only be the case, if the second one contains an IR expression such as `COUNT { MATCH () }`. Therefore, we should not merge two `CREATE` statements if the second one contains an IR expression.

Fixes #13260