Do not merge CREATE clauses if they have dependencies
We merge two consecutive `CREATE` statement to be able to create them together. However, we may not do that if the two have dependencies.
Currently, this can only be the case through an expression referencing a pattern. Therefore, we should not merge two `CREATE` statements if either of them contains an expression referencing a pattern.

Fixes #13260