Block consistency check of diff backups
This never worked since we didn't do the recovery step anyway. Now it is not allowing those backups and printing a reasonable error. It was decided not to allow checking them at all, because it means aggregating the backup chain which can be time-consuming and then in the end it is just thrown away again.