Always use NestedPlanExpressions for SET
SET is currently almost always having row-row-row visibility.
If this will be kept in the long-term is unclear, but for now it should be aligned to always have row-by-row visibility.

The exception was if the SET clause contained a subquery expression that got planned with RollUpApply or similar.
Now, these expressions must always be planned with NestedPlanExpressions, forcing the runtime to evaluate the expressions
interleaved with the set operations.