Refactor tests to actually perform classloading from JARs ()
When testing JAR functionality we are often not actually testing that the classes are loaded by the classloader. This is due to the fact that we in tests are referring to classes that are already loaded by the application classloader. This means that when we load the class from the JARs archive, the application classloader will returned the cached class - not the class from the JAR archive.

This commit changes the behaviour by creating a subclassed type that is not classloaded. This type is written to disk and can then properly be loaded when performing the test.