Adding support for USE clause to admin command planner
This commit does not introduce any new functionality to the product.
It is related to these types of queries:
  USE system SHOW FUNCTIONS
  USE system CALL db.info()
Those queries have been supported since 4.1,
but Fabric/Query router would strip the USE clause from them
before sending them to System DB Cypher stack.
We would like to remove the need to modify AST from Query router,
so we need admin command planner to be tolerant to 'USE system'
in a query.
We need to check that it is really USE system that made it to
admin command planner. The reason is Core API, where no routing is done,
so it must be checked that the user did not do 'USE foo ...' in
a System DB Core API transaction.