Don't use code generation for administration commands ()
Reasoning:
Admin commands are using pipelined runtime internally against the system database. If any of those plans happen to generate permutations of the query string (which could easily be introduced unintentionally with code) it could cause a lot of generated classes for those queries. It might be safer to run these without codegen enabled (except maybe if there are a few commands that are very performance sensitive).